+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 -----
 - Added customer payment sepa support
 - Added specific value image support
+- Added product delivery time manuell and fixed value
 - Changed payment pull error tolerance
 
 1.3.0